Implementing some new things such as search topic, paging, filtering, and sorting

+++ WebApi/WebApi/Controllers/TopicsController.cs
@@ -4,16 +4,20 @@ using Exceptions;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Web;
 using System.Web.Http;
+using System.Web.Http.Routing;
 using Thinktecture.IdentityModel.WebApi;
+using WebApi.Helper;
 
 namespace WebApi.Controllers
 {
     [RoutePrefix("api")]
-    [Authorize]
+    //[Authorize]
     public class TopicsController : SurveyOnlineController
     {
         private ITopicManager _manager { get; set; }
+        private const int MAX_PAGE_SIZE = 10;
 
         public TopicsController(ITopicManager manager)
         {
@@ -76,13 +80,77 @@ namespace WebApi.Controllers
         }
 
         [HttpGet]
-        [ScopeAuthorize("read")]
-        [Route("topics")]
-        public IHttpActionResult GetTopics()
+        //[ScopeAuthorize("read")]
+        [Route("topics", Name = "TopicsList")]
+        public IHttpActionResult GetTopics(string search, string sort = "Id", string category = null,
+            int page = 1, int pageSize = MAX_PAGE_SIZE)
         {
             try
             {
-                return base.Ok(_manager.GetTopics());
+                ICollection<Topic> topics = null;
+
+                if (!string.IsNullOrWhiteSpace(search))
+                {
+                    topics = _manager.GetTopics(search);
+                }
+                else
+                {
+                    topics = _manager.GetTopics();
+                }
+
+                if (topics == null) return Ok(topics);
+
+                var topicList = topics
+                    .AsQueryable<Topic>()
+                    .ApplyFilter(category);
+
+                if (pageSize > MAX_PAGE_SIZE)
+                {
+                    pageSize = MAX_PAGE_SIZE;
+                }
+
+                var totalCount = topicList.Count();
+                var totalPages = (int)Math.Ceiling((double)totalCount / pageSize);
+
+                var urlHelper = new UrlHelper(Request);
+
+                var prevLink = page > 1 ? urlHelper.Link("Topics",
+                    new
+                    {
+                        page = page - 1,
+                        pageSize = pageSize,
+                        sort = sort,
+                        category = category
+                    }) : "";
+                var nextLink = page < totalPages ? urlHelper.Link("TopicsList",
+                    new
+                    {
+                        page = page + 1,
+                        pageSize = pageSize,
+                        sort = sort,
+                        category = category
+                    }) : "";
+
+                var paginationHeader = new
+                {
+                    currentPage = page,
+                    pageSize = pageSize,
+                    totalCount = totalCount,
+                    totalPages = totalPages,
+                    previousLink = prevLink,
+                    nextLink = nextLink
+                };
+
+                HttpContext.Current.Response.Headers.Add("X-Pagination",
+                    Newtonsoft.Json.JsonConvert.SerializeObject(paginationHeader));
+
+                var topicResult = topicList
+                    .ApplySort<Topic>(sort)
+                    .Skip(pageSize * (page - 1))
+                    .Take(pageSize)
+                    .ToList();
+
+                return Ok(topicResult);
             }
             catch (Exception ex)
             {
@@ -94,30 +162,66 @@ namespace WebApi.Controllers
         [HttpPost]
         [ScopeAuthorize("write")]
         [Route("profiles/{userId}/topics")]
-        public IHttpActionResult PostTopic([FromBody] Topic Topic, Guid userId)
+        public IHttpActionResult PostTopic([FromBody] Topic topic, Guid userId)
         {
             try
             {
                 if (!userId.Equals(GetProfileId())) return Unauthorized();
 
-                if (Topic == null) return BadRequest();
+                if (topic == null) return BadRequest();
 
-                var result = _manager.RegisterTopic(Topic);
+                var result = _manager.RegisterTopic(topic);
 
                 if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Created)
-                    return Created(Request.RequestUri + "/" + result.Entity.Id.ToString(), result.Entity);
-
-                if (result.Exception != null)
                 {
-                    if (result.Exception.Code == 1)
+                    if (topic.Categories != null)
                     {
-                        return InternalServerError();
+                        var resultCategories = _manager.RegisterCategories(result.Entity.Id, topic.Categories);
+
+                        if (resultCategories.Status == CoreApi.ActionResult.ManagerActionStatus.Error)
+                        {
+                            return BadRequest(string.Format("El usuario se creo exitosamente pero {0}",
+                                resultCategories.Exception.AppMessage.Message));
+                        }
+                        result.Entity.Categories = resultCategories.Entity;
                     }
-                    else
-                    {
-                        return BadRequest(result.Exception.AppMessage.Message);
-                    }
+
+                    return Created(Request.RequestUri + "/" + result.Entity.Id.ToString(), result.Entity);
                 }
+
+                if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Error)
+                {
+                    return BadRequest(result.Exception.AppMessage.Message);
+                }
+
+                return BadRequest();
+            }
+            catch (Exception ex)
+            {
+                ExceptionManager.GetInstance().Process(ex);
+                return InternalServerError();
+            }
+        }
+
+        [HttpPost]
+        [ScopeAuthorize("write")]
+        [Route("topics/{topicId}/categories")]
+        public IHttpActionResult PostTopicCategories(Guid topicId, ICollection<Category> categories)
+        {
+            try
+            {
+                if (categories == null) return BadRequest();
+
+                var result = _manager.RegisterCategories(topicId, categories);
+
+                if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Created)
+                    return Created(Request.RequestUri, result.Entity);
+
+                if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Error)
+                {
+                    return BadRequest(result.Exception.AppMessage.Message);
+                }
+
                 return BadRequest();
             }
             catch (Exception ex)
@@ -148,7 +252,7 @@ namespace WebApi.Controllers
                     return Ok(result.Entity);
 
                 if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Error)
-                    return InternalServerError();
+                    return BadRequest(string.Concat("El tema pudo ser actualizado pero {0}", result.Exception.AppMessage.Message));
 
                 return BadRequest();
             }
@@ -166,6 +270,8 @@ namespace WebApi.Controllers
         {
             try
             {
+                if (userId.Equals(GetProfileId())) return Unauthorized();
+
                 var result = _manager.DeleteTopic(id, userId);
 
                 if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Deleted)
@@ -173,9 +279,6 @@ namespace WebApi.Controllers
 
                 if (result.Status == CoreApi.ActionResult.ManagerActionStatus.NotFound)
                     return NotFound();
-
-                if (result.Status == CoreApi.ActionResult.ManagerActionStatus.Error && result.Exception != null)
-                    return InternalServerError();
 
                 return BadRequest();
             }
